+++ src/components/ui/personal-info/personal-info.module.scss
